🚚 move: SearchAreaをfeatureに移動
