+++ src/components/feature/SearchArea/SearchBar/index.spec.tsx
