+++ src/components/feature/SearchArea/SearchBar/index.stories.tsx
