+++ src/components/feature/SearchArea/SearchBar/index.tsx
