+++ src/components/feature/SearchArea/SearchFilter/CategoryListbox/index.test.tsx
