+++ src/components/feature/SearchArea/SearchFilter/index.test.tsx
