+++ src/components/feature/SearchArea/index.tsx
